Create index.php
This index.php file displays the functionality of the javascript for the canvas as it runs independently of bootstrap and the (other?) tools used for creating the website layout. If you run this index.php in its own directory on your server, you can see that the canvas drawing function works correctly (the canvas is outlined in black). This index.php file is the same file as paint.php int the public_html/pages folder with the addition of the inline javascript and css which I have added. I show where the code I have added begins and ends with comments. If you are to copy this whole index.php file and paste it into the paint.php file found in the public_html/pages folder, you can see that the functionality is quite different. Likely you will find that clicking and dragging your cursor in the upper-left hand corner of the canvas now results in the corresponding lines being drawn in the lower right-hand part of the canvas. The lines also appear to be much greater in size as if the canvas has been expanded.